requirements.txt
pypdf
PyMuPDF
PyPDF2
Pillow
langchain
clarifai
streamlit